component layout update
